requirements.txt
PyQt6>=6.6.0
PyYAML>=6.0.1
pyperclip>=1.8.2
openai>=1.0.0
pyinstaller
Pillow>=10.0.0
loguru>=0.7.0
google-genai>=0.3.0
